CF template to create an AWS VPC
This is an AWS CloudFormation template to create an AWS VPC.  The
template takes a few input parameters, requiring the user to specify
the AWS region, and the VPC “environment type” (the type of VPC) which
can be admin, dev, stage, or prod.  Everything else is handled by the
template.  The VPC will span three availability zones in the region,
and have a public and private subnet in each availability zone.  A NAT
will be created in each public subnet, and a private routing table will
be created for each private subnet, that uses the NAT in its
availability zone.